requirements.txt
flask==3.0.0
numpy==1.24.3
scikit-learn==1.3.0
anthropic==0.43.0
python-json-logger==2.0.7
gunicorn==21.2.0
python-dotenv==1.0.0
SQLAlchemy~=2.0.36
joblib~=1.4.2
pytz~=2024.2
sentence-transformers~=3.2.1
psycopg2-binary
pandas~=2.2.3
pytest~=8.3.4
jsonschema~=4.23.0
matplotlib~=3.9.3
rdflib~=7.1.1
networkx~=3.2.1
diskcache~=5.6.3

